project/requirements.txt
python-telegram-bot==20.7
Pillow==10.1.0
requests==2.31.0
tenacity==8.2.3
PyYAML==6.0.1 